--- a/TheSadRogue.Primitives/Area.cs
+++ b/TheSadRogue.Primitives/Area.cs
@@ -418,7 +418,7 @@ namespace SadRogue.Primitives
         /// </summary>
         /// <param name="rectangle">Rectangle containing positions to remove.</param>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Remove(Rectangle rectangle) => Remove(rectangle.Positions());
+        public void Remove(Rectangle rectangle) => Remove(rectangle.Positions().ToEnumerable());
 
         /// <summary>
         /// Returns the string of each position in the area, in a square-bracket enclosed list,