--- a/TheSadRogue.Primitives/GridViews/ArrayView2D.cs
+++ b/TheSadRogue.Primitives/GridViews/ArrayView2D.cs
@@ -104,11 +104,9 @@ namespace SadRogue.Primitives.GridViews
         /// <returns/>
         public static ArrayView2D<T> FromMultidimensionalArray(T[,] array) => new ArrayView2D<T>(array);
 
-        /// <summary>
-        /// Sets each element in the ArrayView to the default for type T.
-        /// </summary>
+        /// <inheritdoc/>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Clear() => Array.Clear(_array, 0, _array.Length);
+        public override void Clear() => Array.Clear(_array, 0, _array.Length);
 
         /// <summary>
         /// Returns a string representation of the grid values.