… TheSadRogue.Primitives/Rectangle.cs
… TheSadRogue.Primitives/TheSadRogue.Primitives.csproj